# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Clinical API
  description: This is a RESTful API that efficiently organizes and manages all information and activities related to patients and doctors, improving communication and care.
  contact:
    name: Abdul Muhith
    url: https://portfolio-with-html-css.netlify.app/
    email: amuhith.mern@outlook.com
servers:
  - description: Dev
    url: http://localhost:4000/api/v1
  - description: Prod
    url: https://clinical-api-jillur-m.onrender.com/api/v1
...
paths:
...
  /members/local:
    get:
      operationId: retrieveAllMembers
      tags: [Member]
      summary: Retrieve all members with advanced options.
      description:  This endpoint allows users to retrieve a list of all members in the system, including records for Administrators, Doctors, Nurses, Users or more as necessary. It offers advanced options for sorting, filtering, searching, pagination, along with the ability to specify expanded and selected fields in the query parameters.
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/sizeParam"
        - $ref: "#/components/parameters/sortParam"
        - $ref: "#/components/parameters/sortByParam"
        - $ref: "#/components/parameters/sortOrderParam"
        - $ref: "#/components/parameters/filterParam"
        - $ref: "#/components/parameters/searchParam"
        - $ref: "#/components/parameters/searchTermParam"
        - $ref: "#/components/parameters/expandParam"
        - $ref: "#/components/parameters/selectParam"
      responses:
        200:
          description: Retrieve was successful.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      code:
                        type: integer
                        format: int32
                        example: 200
                      message:
                        type: string
                        description: A human-readable description of the action that occurred.
                        example: A list of members has been retrieved successfully.
                      data:
                        type: array
                        items:
                          allOf:
                            - type: object
                            - description: A detailed list of the specific credentials may be utilized, along with the both data and meta for pagination and hate-OAS links.
                            - $ref: "#/components/schemas/UserDTO"
                            - $ref: "#/components/schemas/Timestamp"
                            - type: object
                              properties:
                                link:
                                  type: string
                                  format: url
                                  example: /api/v1/local/users/60d21b4667d0d8992e610c88
                      meta:
                        type: object
                        properties:
                          pagination:
                            $ref: "#/components/schemas/pagination"
                          links:
                            type: object
                            properties:
                              self:
                                type: string
                                example: /api/v1/members/local?page=2&size=25
                              nextPage:
                                type: string
                                example: /api/v1/members/local?page=3&size=25
                              prevPage:
                                type: string
                                example: /api/v1/members/local?page=1&size=25
                              administrators:
                                type: string
                                example: /api/v1/administrators
                              doctors:
                                type: string
                                example: /api/v1/doctors/local
                              nurses:
                                type: string
                                example: /api/v1/nurses/local
                              users:
                                type: string
                                example: /api/v1/users/local
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 400
                  message:
                    type: string
                    description: A human-readable description of the error that occurred.
                    example: Validation Error
                  errors:
                    allOf:
                      - type: array
                        description: An array detailing specific errors encountered in the request.
                        items:
                          $ref: "#/components/schemas/BadRequestError"
                      - $ref: "#/components/schemas/RetrieveAllItemsErrorResponse"
                  hints:
                    type: string
                    description: Suggestions or tips to help resolve the error.
                    example: Please, ensure all query parameters are correctly filled and try again. If the issue persists, feel free to contact our technical support.
                  trace_id:
                    $ref: "#/components/schemas/TraceID"
        404:
          description: Resource Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 404
                  message:
                    type: string
                    description: A human-readable description of the error that occurred.
                    example: Invalid Credentials
                  errors:
                    allOf:
                      - type: array
                        description: An array detailing specific errors encountered in the request.
                        items:
                          $ref: "#/components/schemas/NotFoundError"
                      - example:
                        - code: RESOURCE_NOT_FOUND
                          message: The resource you are looking for could not be found.
                          field: null
                          location: query
                          status: 404
                  hints:
                    type: string
                    description: Suggestions or tips to help resolve the error.
                    example: Please, ensure all query parameters are correctly filled and try again. If you continue to have issues, please reach out to our support team.
                  trace_id:
                    $ref: "#/components/schemas/TraceID"
        500:
          $ref: "#/components/responses/500"
        503:
          $ref: "#/components/responses/503"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
          example: 2024-09-21T12:34:56Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-09-21T12:34:56Z
    UserDTO:
      allOf:
        - type: object
          properties:
            id:
              type: string
              description: Must be MongoDB Object Id
              example: 60d21b4667d0d8992e610c88
            username:
              type: string
              example: Abdullah
            phone:
              type: string
              example: "+8801596528545"
            email:
              type: string
              format: email
              example: abdullah@example.com
            role:
              type: string
              enum: [user, nurse, doctor, administrator]
              example: "user"
            status:
              type: string
              enum: [active, disabled, request changed, pending]
              example: pending
    pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        size:
          type: integer
          format: int32
          example: 25
        nextPage:
          type: integer
          format: int32
          example: 3
        prevPage:
          type: integer
          format: int32
          example: 1
        totalPages:
          type: integer
          format: int32
          example: 5
        totalItems:
          type: integer
          format: int32
          example: 52
...
    TraceID:
      type: string
      format: uuid
      description: A unique identifier for the error, useful for tracking and debugging within the system.
      example: 550e8400-e29b-41d4-a716-446655440000
    BadRequestError:
      type: object
      properties:
        code:
          type: string
          description: A short code representing the type of error.
        message:
          type: string
          description: A detailed message describing the specific error.
        field:
          type: string
          description: The name of the field that caused the error (if applicable).
        location:
          type: string
          enum: [body, query, path, null]
          description: The location of the error within the request (e.g., "body" or "query").
        status:
          type: integer
          format: int32
          description: An HTTP status code associated with the specific error, usually 400.
    NotFoundError:
      type: object
      properties:
        code:
          type: string
          description: A short code representing the type of error.
        message:
          type: string
          description: A detailed message describing the specific error.
        field:
          type: string
          description: The name of the field that caused the error (if applicable).
        location:
          type: string
          enum: [body, query, path, null]
          description: The location of the error within the request (e.g., "body" or "query").
        status:
          type: integer
          format: int32
          description: An HTTP status code associated with the specific error, usually 404.
    InternalServerError:
      type: object
      properties:
        code:
          type: string
          description: A short code representing the type of error.
        message:
          type: string
          description: A detailed message describing the specific error.
        field:
          type: string
          description: The name of the field that caused the error (if applicable).
        location:
          type: string
          enum: [server, null]
          description: The location of the error within the request (e.g., "server" or "null").
        status:
          type: integer
          format: int32
          description: An HTTP status code associated with the specific error, usually 500.
...
    RetrieveAllItemsErrorResponse:
      example:
        - code: INVALID_PAGE
          message: The page parameter can't be a negative number.
          field: page
          location: query
          status: 400
        - code: INVALID_SIZE
          message: The size parameter should be a number between 1 and 100.
          field: size
          location: query
          status: 400
        - code: INVALID_SORT
          message: The sort parameter format is invalid.
          field: sort
          location: query
          status: 400
        - code: INVALID_SORT_BY
          message: The sort-by parameter provided is invalid.
          field: sort_by
          location: query
          status: 400
        - code: INVALID_SORT_ORDER
          message: The sort-order parameter should be either ascending or descending.
          field: sort_order
          location: query
          status: 400
        - code: INVALID_FILTER
          message: The filter parameter should be a key-value format.
          field: filter
          location: query
          status: 400
        - code: INVALID_SEARCH
          message: The search parameter should be a key-value format.
          field: search
          location: query
          status: 400
        - code: INVALID_SEARCH_TERM
          message: The search-term parameter provided is invalid.
          field: search_term
          location: query
          status: 400
        - code: INVALID_EXPAND
          message: The expand parameter format is invalid.
          field: expand
          location: query
          status: 400
        - code: INVALID_SELECT
          message: The select parameter provided is invalid.
          field: select
          location: query
          status: 400
        - code: INVALID_QUERY
          message: One or more query parameters are invalid.
          field: page, size, sort, sort_by, sort_order, filter, search, search_term, expand and select
          location: query
          status: 400
...
  responses:
...
    500:
      description: Server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              message:
                type: string
                description: A human-readable description of the error that occurred.
                example: Internal Server Error
              errors:
                type: array
                description: An array detailing specific errors encountered in the request.
                items:
                  $ref: "#/components/schemas/InternalServerError"
                example:
                  - code: SERVER_ERROR
                    message: An unexpected error occurred on the server.
                    field: null
                    location: server
                    status: 500
              hints:
                type: string
                description: Suggestions or tips to help resolve the error.
                example: We are sorry for the inconvenience. Please try again later or reach out to support for help.
              trace_id:
                $ref: "#/components/schemas/TraceID"
    503:
      description: Service Unavailable
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 503
              message:
                type: string
                description: A human-readable description of the error that occurred.
                example:  Service is unavailable
              errors:
                type: array
                description: An array detailing specific errors encountered in the request.
                items:
                    $ref: "#/components/schemas/InternalServerError"
                example:
                  - code: SERVICE_UNAVAILABLE
                    message: The service is currently (2024-10-12T12:34:56Z) unavailable.
                    field: null
                    location: server
                    status: 503
              hints:
                type: string
                description: Suggestions or tips to help resolve the error.
                example: We are sorry for the inconvenience. The service is temporarily down for maintenance. Please try again shortly or reach out to support for help.
              trace_id:
                $ref: "#/components/schemas/TraceID"
  parameters:
...
    pageParam:
      name: page
      in: query
      required: false
      description: The page number to retrieve when results are paginated in conjunction with the `size` parameter. If not specified, the default will be the first page of results.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    sizeParam:
      name: size
      in: query
      required: false
      description: The number of items to retrieve per page, used in conjunction with the `page` parameter. If not specified, the default is `25` items per page, with a maximum allowed value of `100`.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 25
    sortParam:
      name: sort
      in: query
      required: false
      description: A comma-separated list of fields to sort the results by, along with their sort order. Prefix a field name with a hyphen (`-`) to indicate descending order. If not provided, results will default to ascending order.
      schema:
        type: string
    sortByParam:
      name: sort_by
      in: query
      required: false
      description: A comma-separated list of fields used for sorting in conjunction with the `sort_order` parameter. If not provided, results will default based on use case.
      schema:
        type: string
    sortOrderParam:
      name: sort_order
      in: query
      required: false
      description: This is used in conjunction with the `sort_by` parameter for specifying whether the sort is ascending or descending. If not provided, results will default to ascending order.
      schema:
        type: string
        enum: [ascending, descending]
        default: ascending
    filterParam:
      name: filter
      in: query
      required: false
      description: A comma-separated list of expressions used to narrow down the results based on specific criteria, where each filter expression is formatted as a key-value pair. If not provided, the results will default to include all available data without any filtering.
      schema:
        type: string
        format: key-value
        default: ""
    searchParam:
      name: search
      in: query
      required: false
      description: This is a field-specific search that follows a key-value format to find resources by matching against their attributes. If not provided, the default setting will ensure that no related resources are included in the search results.
      schema:
        type: string
        format: key-value
        default: ""
    searchTermParam:
      name: search_term
      in: query
      required: false
      description: This is a free-text searching by providing a keyword to search for matches in all relevant fields. If not provided, the default setting will ensure that no related resources are included in the search results.
      schema:
        type: string
        default: ""
    expandParam:
      name: expand
      in: query
      required: false
      description: A comma-separated list of fields to expand, where each field may specifies additional subfields to include in the result using `dot notation`. If not specified, all related resource fields will be expanded. Typically, results will default based on use case.
      schema:
        type: string
        default: ""
    selectParam:
      name: select
      in: query
      required: false
      description: A comma-separated list of fields that specifies additional fields to include in the result. If not provided, results will default based on use case.
      schema:
        type: string
        default: ""